add app folder for the app module
