+++ src/app/dto/runDiagramDTO.ts
@@ -1,0 +1,5 @@
+import { SerializedDiagram } from '@data-story-org/core';
+
+export class RunDiagramDTO {
+  model: SerializedDiagram;
+}
